Add initial presentation pages.
I want to try git pages because 1) I do not like powerpoint, and 2) I want
to include at least one d3.js example and I do not know if powerpoint can
handle that. I guess if this does not work that is still another option.

I could put these in separate pages, but I am not sure how much support git
pages has for that type of thing and then there would also be issues with
keeping the template similar. So instead I ran with splitting the slides up
using JavaScript.

If this works it will be awesome... if it does not, well at least I will have
plain HTML pages to show.